Fix unnecessary refresh of notes list when saving a note from another folder. In KntServerCOMCtrl, parameterize RetroDelay.

+++ Src/KNote/ClientWin/Controllers/NotesSelectorCtrl.cs
@@ -180,11 +180,14 @@ public class NotesSelectorCtrl : CtrlSelectorBase<IViewSelector<NoteInfoDto>, No
                     ListEntities.Add(note);
                     View.AddItem(note);
                 }
-            }                
+            }
             else
             {
-                ListEntities.RemoveAll(_ => _.NoteId == note.NoteId);
-                View.DeleteItem(note);
+                if(updateNote != null)
+                {
+                    ListEntities.RemoveAll(_ => _.NoteId == note.NoteId);
+                    View.DeleteItem(note);
+                }
             }
         }
     }
